Finished simulateSale()
- Program can now simulate a sale, including changing the amount of contents in storage bins and making a transaction object.

CoffeeTruck()
- Small edits in returnMenu(), just changed integer div to float div
- created isDrinkAvailable. Using the output from returnMenu(), and users input (type and size), determines is drink is available.
- simulateSale() is not fully functional.

!!!!!!!!
In the even that storageBin1 has 10grams of coffee, and storageBin2 has 10 grams of coffee, and the drink requires 20 grams of coffee, the drink is still considered available. However, in the current program theres a missing if-statement for this case. Comment is in line 284 for this to be added.

Ingredient.java
- Edited default values of items lol. Original values were very overpriced

JavaJeep
- Added a clear() function [doesnt actually clear screen, just prints a bunch of blank newlines]

Transaction.java
- Transaction can now be initialized and creates the correct values for all its attributes.
- Added getIngredients(). returns ingredients arraylist